Initial SDL2 work

+++ src/xmoto/Sound.cpp
@@ -116,7 +116,7 @@ bool Sound::isInitialized() {
 
 void Sound::update(void) {}
 
-int Sound::RWops_seek(SDL_RWops *context, int offset, int whence) {
+int64_t Sound::RWops_seek(SDL_RWops *context, int64_t offset, int whence) {
   FileHandle *pf = (FileHandle *)context->hidden.unknown.data1;
   switch (whence) {
     case SEEK_SET:
@@ -132,7 +132,7 @@ int Sound::RWops_seek(SDL_RWops *context, int offset, int whence) {
   return XMFS::getOffset(pf);
 }
 
-int Sound::RWops_read(SDL_RWops *context, void *ptr, int size, int maxnum) {
+size_t Sound::RWops_read(SDL_RWops *context, void *ptr, size_t size, size_t maxnum) {
   FileHandle *pf = (FileHandle *)context->hidden.unknown.data1;
   if (XMFS::isEnd(pf))
     return 0;
@@ -147,7 +147,7 @@ int Sound::RWops_read(SDL_RWops *context, void *ptr, int size, int maxnum) {
   return nToRead;
 }
 
-int Sound::RWops_write(SDL_RWops *context, const void *ptr, int size, int num) {
+size_t Sound::RWops_write(SDL_RWops *context, const void *ptr, size_t size, size_t num) {
   return num;
 }
